refactor: Move rule schemas to a new file and update imports
